댓글 기능 추가완료

+++ BE/src/main/java/wealthwise/BE/controller/CommentController.java
@@ -22,11 +22,13 @@ public class CommentController {
      * @param auth 사용자 인증 정보
      * @return 생성된 댓글의 ID
      */
-
     @PostMapping("/{boardId}")
     public Long addComment(@PathVariable Long boardId,
                            @RequestBody CommentCreateRequest req,
                            Authentication auth) {
+        if (auth == null) {
+            throw new RuntimeException("로그인이 필요합니다.");
+        }
         return commentService.writeComment(boardId, req, auth.getName());
     }
 
@@ -42,6 +44,9 @@ public class CommentController {
     public Long editComment(@PathVariable Long commentId,
                             @RequestBody CommentCreateRequest req,
                             Authentication auth) {
+        if (auth == null) {
+            throw new RuntimeException("로그인이 필요합니다.");
+        }
         Long boardId = commentService.editComment(commentId, req.getBody(), auth.getName());
         return boardId;
     }
@@ -55,6 +60,9 @@ public class CommentController {
      */
     @DeleteMapping("/{commentId}")
     public Long deleteComment(@PathVariable Long commentId, Authentication auth) {
+        if (auth == null) {
+            throw new RuntimeException("로그인이 필요합니다.");
+        }
         Long boardId = commentService.deleteComment(commentId, auth.getName());
         return boardId;
     }
